Fix license
The MIT license requires a "+ file LICENSE" in the DESCRIPTION file. It 
must be readable using read.dcf or it results in an R CMD check note. If 
the LICENSE file is supplied as such, it consists of only two lines and 
the MIT license doesn't show up on GitHub, which isn't optimal. Using 
usethis::use_mit_license(), there is also a LICENSE.md, but this scheme 
shows up as two licenses in GitHub and the LICENSE file which doesn't 
mention MIT takes precedence. LICENSE shows up as unknown, and there 
doesn't seem to be a way to not deny the file from showing up as a 
license. LICENSE.md doesn't show the "fact box" regarding the MIT 
license in GitHub, probably because of the .md extension. I did it as 
per R CMD check for now, there are some open issues open regarding 
functionality for denying unknown LICENSES from showing up on GitHub. 
The license tab at the right shows that it is licensed under MIT 
anyways.