Harmonize documentation
Harmonized vignettes and documentation; the conceptualization of the 
package presented in the vignette carries over to notame-package.Rd, 
where the descriptions of functions now match the title in the 
function-specific .Rd files. Also changed name of clean_stats_results to 
summarize_results and added the "Preprocessing" BiocViews keyword to 
resolve a note.