Fix seed issue
In notame 0.3.2, a seed is set repeatedly for visualizations() and 
perform_lmer(). Earlier I simply removed these seeds. Now, for 
visualizations(), I did a workaround so the global random number stream 
is not affected plus there is no warning from set.seed for the 
submission. 

In perform_lmer, bootstrap confidence intervals are reproducible by 
registering the BiocParallel backend with the RNGseed argument, which I 
mentioned in the details. The confidence intervals aren't reproducible 
on a feature-to-feature basis but a run-to-run basis, again to avoid 
set.seed(). This is perhaps not optimal, let's change it if it is deemed 
a problem.