Almost completely updated
Almost completely updated for the new example data.

The project_example vignette could use more work (batch correction or 
using a single batch, for example) to make plots more typical. 

I'm struggling updating the Mann-Whitney/Wilcoxon signed rank test. The 
estimates largely have different signs from the manually computed 
medians. I believe the manual part is based on that if distributions of 
groups are equal, Mann-Whitney/Wilcoxon signed rank test estimates are 
the same as the median or mean. This may somehow not be the case in the 
new example data, but I can't quite figure it out. I have the same 
problem using the new data with notame 0.3.2, so the problem doesn't 
seem to be with the new perform_non_parametric."