Fix final unit tests
I found that the Mann-Whitney unit test sign equality logic is sensitive 
to which seed was used in generating the example data. I removed that 
part of the unit test, since it clearly isn't working as expected, it 
didn't work with a different seed in notame 0.3.2 either. The test still 
tests the statistic and I deemed this enough.

Regarding Wilcoxon signed rank, I noticed that I had specified 
Spearman's correlation although the default (Pearson) was used in 0.3.2. 
This resolved the issue I was having where the effect between time 
points/batches results in all signs equal to 0, throwing a zero-variance 
error. The original Pearson works though, although I didn't fully 
explore the details and logic.