Initial commit
Tasked with a a SummarizedExperiment translation of notame, I thought it 
was best to include an input check function. The input check function 
check_object() checks the object as if one was constructing a MetaboSet 
object, so no additional input checks need to be written. This is not 
optimal for interoperability with other SummarizedExperiment 
functionality as the user may have to fix things which are not 
necessarily needed in the specific function. This problem is, in part, 
alleviated using fix_object(), which does the modifications which 
formerly were included in check_feature_data, check_pheno_data, 
check_exprs etc. The division into check_object() and fix_object() 
should make it clear that check_object() doesn't do any changes, merely 
checks, while fix_object() does changes. Going forward, we may consider 
removing some parts of the input checks and fixes and inserting them 
directly as input checks to functions for more interoperability.

To keep track of the logic, I also updated some old functions and 
validity to use the new check functions. The tests needed to be updated 
only minimally. Tests pass, examples pass, there seems to be only small 
issues relating to documentation and namespace.