Update functionality for SE
The code has been adapted to SummarizedExperiment. Included setAs method 
for conversions as instructed by Bioconductor. Everything works (tested 
thoroughly also in contrast to original notame) although it doesn't pass 
checks yet as some things remain to be resolved before updating 
documentation and finishing off. I also aimed to make notame 
interoperable as in avoiding unexpected problems because of naming and 
structure of the SummarizedExperiment by using check_object() and 
fix_object().

Some changes of note:
1. coercion from S4Vectors::DataFrame to data.frame
- we could consider tidySummarizedExperiment dependency for this
2. the use of colnames and rownames instead of samplesNames and 
featureNames (and their respective notame-methods)
- this is warranted in that the input checks must check for identical 
object$Feature_ID and rownames(object)
3. changes in transpositions
- especially invoving pcaMethods

To-do
- Documentation
- Unit tests
- Since we're using SummarizedExperiment which doesn't have the special 
slots, the show method for MetaboSet couldn't be translated directly. 
Moreover, I could imagine that it isn't good practice to define a new 
show method. Addressing this.
- I still need to work on input checks for interoperability, but I 
figured this is not critical as long as it works as is should, which it 
does now when constrained by check_object(). I also need to specify some 
default values for parameters which originally were assigned one of the 
special MetaboSet slots.