Fix example data (#3)
* Preparation for the following plan; examples, unit tests and removal of some data done as if we decided to follow through with the plan below (for illustration). Vignette left for now along with sample_data_whole.xlsx so that package builds. - remove merged_sample.rda, sample_data_whole.xlsx, old mode-specific .xlsx files, - modify example_data.R to accommodate all examples, where we first create modes separately and save as column_mode_sample.rda, then merge them and save as example_set.rda. We also save mode-specific .xlsx files (commented out). - easy_data.xlsx, split_data.xlsx included under the testthat folder instead

* Cleaning

Keeping in sync with refactor-simple-tests for checks, for example

* Modify data and update other aspects accordingly

Package data modified and re-saved. The examples and tests updated 
accordingly, except for a few odds and ends, and the plots look quite 
representative after normalization. Also removed a few remaining 
merged_sample uses/mentions. There is no systematic difference between 
groups, only between time points. If this looks good overall, I'll go 
ahead and fix the remaining issues and otherwise prepare for merge.

* Almost completely updated

Almost completely updated for the new example data.

The project_example vignette could use more work (batch correction or 
using a single batch, for example) to make plots more typical. 

I'm struggling updating the Mann-Whitney/Wilcoxon signed rank test. The 
estimates largely have different signs from the manually computed 
medians. I believe the manual part is based on that if distributions of 
groups are equal, Mann-Whitney/Wilcoxon signed rank test estimates are 
the same as the median or mean. This may somehow not be the case in the 
new example data, but I can't quite figure it out. I have the same 
problem using the new data with notame 0.3.2, so the problem doesn't 
seem to be with the new perform_non_parametric."

* Fix final unit tests

I found that the Mann-Whitney unit test sign equality logic is sensitive 
to which seed was used in generating the example data. I removed that 
part of the unit test, since it clearly isn't working as expected, it 
didn't work with a different seed in notame 0.3.2 either. The test still 
tests the statistic and I deemed this enough.

Regarding Wilcoxon signed rank, I noticed that I had specified 
Spearman's correlation although the default (Pearson) was used in 0.3.2. 
This resolved the issue I was having where the effect between time 
points/batches results in all signs equal to 0, throwing a zero-variance 
error. The original Pearson works though, although I didn't fully 
explore the details and logic.