Make input checks more specific
The logic for the input checks is as follows. It was not clear to what 
extent the functions largely relied on the checks in read_from_excel, 
construct_metabosets and setValidity for MetaboSet, with only tens of 
input checks and type conversions throughout the rest of the package. 
The input checks in the above functions are now used through 
check_object with optional arguments, which is used for fine-grained 
input checks in every function. This also means that error messages 
relating to the input from object are consistent and easily updated. 
Informative downstream error messages (many from ggplot2, for example) 
were not "overwritten".

Also ensured consistent logic in returning MetaboSet and/or 
SummarizedExperiment. assess_quality and drop_flagged were made into 
methods as they are exported and used internally, which complicated the 
logic in handling SE and MetaboSet.

To-do: 
- checking and returning specific assay of SE in check_object
- settle on behavior where default values used to be group_col, time_col 
or subject_col
- a bit more work on input checks, for example for formula interface in 
stats, consider skipping some hard-coded ones for interoperability