Support for multiple assays
notame now supports multiple assays. If no the object contains only one 
assay, the assay.type parameter and name parameter does not need to be 
specified, so the interface didn't change compared to MetaboSet. 

Examples and tests pass, but need to be updated to use multiple assays 
for demonstration and coverage. Then there's a few odds and ends such as 
default values (# CONSIDER) documentation and clearning.