Unit tests
Wrote unit tests to cover new functionality. In many cases, unit tests 
for shallow wrappers such as .check_object() would have overlapped so 
much with existing unit tests that I did not make separate unit tests. 
Instead, I focused on important cases for assay control and MetaboSet 
support. Testing could definitely be improved, for example for checking 
helpers such as .check_pheno_data, but this would have been more of an 
improvement on the pre-existing tests, so this is better done separately 
to be systematic. 

Passes checks, there are a few notes not related to this 
SummarizedExperiment translation.

Special cases:
1. The default behavior in merge_objects is, as of now, motivated in 
that with an arbitrary number of merges, it is not convenient to supply 
assay.type anyways and after a merge, one is ready for statistical 
analysis so dropping the other assays is fine, methinks.  

2. perform_correlation_tests() is consistent and simple in that one 
wouldn't typically perform correlation tests between different versions 
of the same assay in the object. Could use altExp in 
TreeSummarizedExperiment, tho.

3. align_batches() was not modified for selecting both assays assay from 
an object now as we're currently working on batchCorr, too, and this is 
simple and works fine.

4. save_batch_plots() was modified to accept several assays from a 
single object.

One shortcoming is the log methods, where making SummarizedExperiment 
methods using assay.type would overwrite the generic. For present 
purposes, I simply kept the log methods as is (now overwriting generic), 
although they currently don't work when using multiple assays so this is 
important to resolve. On the other hand, this is in line with the 
vignettes and examples, which I didn't change at this point as they now 
work for both SummarizedExperiment and MetaboSet.