Address immediately applicable feedback
I didn't consider batchCorr-functionality at this stage as it will be 
commented out/removed in the submission. 

For the perform_t_test/perform_non_parametric issue, I could have 
reverted it to apply the parallel looping similarly to other 
paralellized univariate functionality, but using mapply addresses the 
performance issues on Windows (of late summer) as we can pass only the 
data needed for each iteration and the feature name. The performance is 
around 10x on my computer, so that may be considered for other 
paralellized functionality, too, although the performance gain is 
smaller for computations that take a longer time.

I had mistakenly thought that BiocParallel doesn't propagate message()s, 
so I could now simplify error handling (this is important because 
parallel mode is default when using BiocParallel). Was also able to 
revert one of the tests because of simpler error handling and using 
dplyr::bind_rows. 

Other notes (also regarding review comments)
- both SummarizedExperiment and MetaboSet seems to be mentioned in all 
exported documentation
- batchCorr is now hosted on https://github.com/MetaboComp/batchCorr
- SummarizedExperiment-support for batchCorr will provide near identical 
functionality to old batchCorr-dependent functionality in notame
- DelayedArray works save for a few exceptions, I have
notes and a working solution for this if needed
- continuing resolving issues in new branches
- unit tests: My approach until this point was to exhaustively test the 
functionality with the same examples and additional examples in parallel 
with the original notame version. I have not done this exhaustively from 
a coverage point of view, but meticulously and comparing the results of 
analysis scripts. I'm quite confident in the functionality as is. I aim 
to do more test-driven development and this would definitely support a 
larger reworking of the package. Lower priority.
- proper linking in docs: Postponing this until during submission or 
initial feedback in case major restructuring is needed
- update parameter names: Lower priority. Should we have "_" and "." 
naming conventions simultaneously? Context: there should be no 
deprecations in the submission
- update function names? Should we include two sets of functions 
simultaneously? Context: no deprecations/defunct
- copyright: Artur checked the copyright thing with Kati changed it 
accordingly (I'll double-check this)