Fix data reading bug
Resolved the issue of re-created Feature_ID column in read_from_excel. 
The tests updated in this commit were introduced in this branch, so 
basically assuring original behavior: if a column x exists, don't modify 
it when using fix_object. This behavior is consistent except for 
Sample_ID column. Clarified documentation a bit.