Fix examples with long run time
Doing less repeats of cross-validation in mixOmics wrappers. Also 
included nrepeat parameter in call to mixOmics::perf(), believing there 
was formerly a mistake here with 50 repeats hard-coded. Only need to 
shave off seconds or so for R CMD check (max 10 minutes), here goes a 
minute or so. Further reduction of check time is not trivial.